Working app

+++ src/main/java/com/example/Product.java
@@ -4,10 +4,10 @@ public class Product {
 
     public String name;
     public double price;
-    public int quantity;
+    public double quantity;
     public String unit;
 
-    public Product(String name, double price, int quantity, String unit) {
+    public Product(String name, double price, double quantity, String unit) {
         this.name = name;
         this.price = price;
         this.quantity = quantity;
